Add files via upload
Quick template demonstrating variable initialization, declaration along with pointer initialization and declaration. This is to help me remember the syntax and the theory behind it.

The printf functions demonstrate both the variable 'a' and *p are the same value and the how the address of variable a and the pointer are in the same location in memory.