Add files via upload
More practice with array declaration/initialization along with printing all elements of the array.

Then, using the pointer and re-assigning the pointer to display various indices of the fixed size array.

Will occasionally display the pointer variable 'p' itself and the address in memory it is currently located. This does throw compiler warnings in the online compiler I used.